Reincluded glTransate for Cylinder, changed calling of colour change from inside .cpp to in main
Reincluded glTranslate for cylinder so that it matches the assignment specs, and changed ModelVehicle so that the wheels sit right.
Also, got told in the tute to change where we set our colour from inside each shape's cpp file to outside in the main file. This way you can draw shapes of the same style in different colours,  e.g.  red rectangular prism and then a blue rectangular prism.